services/api-gateway/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.30
boto3>=1.34
python-dotenv>=1.0
pydantic>=2,<3
dramatiq>=1.16
redis>=5.0
